easyowov2/requirements.txt
torch~=2.2.1+cu118

torchvision~=0.17.1+cu118

opencv-python~=4.8.1.78

thop~=0.1.1-2209072238

scipy~=1.10.1

matplotlib~=3.2.2

numpy~=1.24.1

imageio~=2.34.0

pillow~=10.2.0